Create README.md
# Mastermind
Mastermind is a two-player code-breaking game, in which one player hides a code consisting of colors, while the other player has to guess it using clues given by the former player for each turn. in this code computer generate secret then you should to guess the secret. "a" is number of color with correct position "b" is number of color with correct color but wrong position.
first click generate secret then you can play and choose color then click on check. 
when a=4 , b=0 you win.